Remove MonoBehaviour PlayerController
