--- Assets/Scripts/PlayerController.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9ed34737a4e5d2d4ea4fa4eb798800a5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
